Update model class

+++ week3_oop_and_java_basic/payment_system/src/model/PaymentMethod.java
@@ -1,17 +1,19 @@
 package model;
 
+import java.math.BigDecimal;
+
 public abstract class PaymentMethod {
     protected String methodId;
     protected String methodName;
-    protected double balance;
+    protected BigDecimal balance;
 
-    public PaymentMethod(String methodId, String methodName, double balance) {
+    public PaymentMethod(String methodId, String methodName, BigDecimal balance) {
         this.methodId = methodId;
         this.methodName = methodName;
         this.balance = balance;
     }
 
-    public abstract boolean processPayment(double amount, boolean isForeignTransaction);
-    public abstract boolean processRefund(double amount);
+    public abstract boolean processPayment(BigDecimal amount, boolean isForeignTransaction);
+    public abstract boolean processRefund(BigDecimal amount);
 
 }
